some corrections tests
